% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.elastictranscoder_operations.R
\name{create_job}
\alias{create_job}
\title{When you create a job, Elastic Transcoder returns JSON data that includes the values that you specified plus information about the job that is created}
\usage{
create_job(PipelineId, Input = NULL, Inputs = NULL, Output = NULL,
  Outputs = NULL, OutputKeyPrefix = NULL, Playlists = NULL,
  UserMetadata = NULL)
}
\arguments{
\item{PipelineId}{The \code{Id} of the pipeline that you want Elastic Transcoder to use for transcoding. The pipeline determines several settings, including the Amazon S3 bucket from which Elastic Transcoder gets the files to transcode and the bucket into which Elastic Transcoder puts the transcoded files.}

\item{Input}{A section of the request body that provides information about the file that is being transcoded.}

\item{Inputs}{A section of the request body that provides information about the files that are being transcoded.}

\item{Output}{A section of the request body that provides information about the transcoded (target) file. We strongly recommend that you use the \code{Outputs} syntax instead of the \code{Output} syntax.}

\item{Outputs}{A section of the request body that provides information about the transcoded (target) files. We recommend that you use the \code{Outputs} syntax instead of the \code{Output} syntax.}

\item{OutputKeyPrefix}{The value, if any, that you want Elastic Transcoder to prepend to the names of all files that this job creates, including output files, thumbnails, and playlists.}

\item{Playlists}{If you specify a preset in \code{PresetId} for which the value of \code{Container} is fmp4 (Fragmented MP4) or ts (MPEG-TS), Playlists contains information about the master playlists that you want Elastic Transcoder to create.

The maximum number of master playlists in a job is 30.}

\item{UserMetadata}{User-defined metadata that you want to associate with an Elastic Transcoder job. You specify metadata in \code{key/value} pairs, and you can add up to 10 \code{key/value} pairs per job. Elastic Transcoder does not guarantee that \code{key/value} pairs are returned in the same order in which you specify them.}
}
\description{
When you create a job, Elastic Transcoder returns JSON data that includes the values that you specified plus information about the job that is created.
}
\details{
If you have specified more than one output for your jobs (for example, one output for the Kindle Fire and another output for the Apple iPhone 4s), you currently must use the Elastic Transcoder API to list the jobs (as opposed to the AWS Console).
}
\examples{

}
