% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.elastictranscoder_operations.R
\name{create_preset}
\alias{create_preset}
\title{The CreatePreset operation creates a preset with settings that you specify}
\usage{
create_preset(Name, Container, Description = NULL, Video = NULL,
  Audio = NULL, Thumbnails = NULL)
}
\arguments{
\item{Name}{The name of the preset. We recommend that the name be unique within the AWS account, but uniqueness is not enforced.}

\item{Container}{The container type for the output file. Valid values include \code{flac}, \code{flv}, \code{fmp4}, \code{gif}, \code{mp3}, \code{mp4}, \code{mpg}, \code{mxf}, \code{oga}, \code{ogg}, \code{ts}, and \code{webm}.}

\item{Description}{A description of the preset.}

\item{Video}{A section of the request body that specifies the video parameters.}

\item{Audio}{A section of the request body that specifies the audio parameters.}

\item{Thumbnails}{A section of the request body that specifies the thumbnail parameters, if any.}
}
\description{
The CreatePreset operation creates a preset with settings that you specify.
}
\details{
Elastic Transcoder checks the CreatePreset settings to ensure that they meet Elastic Transcoder requirements and to determine whether they comply with H.264 standards. If your settings are not valid for Elastic Transcoder, Elastic Transcoder returns an HTTP 400 response (\code{ValidationException}) and does not create the preset. If the settings are valid for Elastic Transcoder but aren't strictly compliant with the H.264 standard, Elastic Transcoder creates the preset and returns a warning message in the response. This helps you determine whether your settings comply with the H.264 standard while giving you greater flexibility with respect to the video that Elastic Transcoder produces.

Elastic Transcoder uses the H.264 video-compression format. For more information, see the International Telecommunication Union publication \emph{Recommendation ITU-T H.264: Advanced video coding for generic audiovisual services}.
}
\examples{

}
