% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.elastictranscoder_operations.R
\name{delete_pipeline}
\alias{delete_pipeline}
\title{The DeletePipeline operation removes a pipeline}
\usage{
delete_pipeline(
  Id = "string"
)
}
\arguments{
\item{Id}{[required] The identifier of the pipeline that you want to delete.}
}
\description{
The DeletePipeline operation removes a pipeline.
}
\details{
You can only delete a pipeline that has never been used or that is not currently in use (doesn't contain any active jobs). If the pipeline is currently in use, \code{DeletePipeline} returns an error.
}
\examples{

}
