% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.elastictranscoder_operations.R
\name{delete_preset}
\alias{delete_preset}
\title{The DeletePreset operation removes a preset that you've added in an AWS region}
\usage{
delete_preset(Id)
}
\arguments{
\item{Id}{[required] The identifier of the preset for which you want to get detailed information.}
}
\description{
The DeletePreset operation removes a preset that you've added in an AWS region.
}
\details{
You can't delete the default presets that are included with Elastic Transcoder.
}
\section{Accepted Parameters}{
\preformatted{delete_preset(
  Id = "string"
)
}
}

