% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.elastictranscoder_operations.R
\name{list_jobs_by_pipeline}
\alias{list_jobs_by_pipeline}
\title{The ListJobsByPipeline operation gets a list of the jobs currently in a pipeline}
\usage{
list_jobs_by_pipeline(PipelineId, Ascending = NULL, PageToken = NULL)
}
\arguments{
\item{PipelineId}{The ID of the pipeline for which you want to get job information.}

\item{Ascending}{To list jobs in chronological order by the date and time that they were submitted, enter \code{true}. To list jobs in reverse chronological order, enter \code{false}.}

\item{PageToken}{When Elastic Transcoder returns more than one page of results, use \code{pageToken} in subsequent \code{GET} requests to get each successive page of results.}
}
\description{
The ListJobsByPipeline operation gets a list of the jobs currently in a pipeline.
}
\details{
Elastic Transcoder returns all of the jobs currently in the specified pipeline. The response body contains one element for each job that satisfies the search criteria.
}
\examples{

}
