% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.elastictranscoder_operations.R
\name{list_jobs_by_status}
\alias{list_jobs_by_status}
\title{The ListJobsByStatus operation gets a list of jobs that have a specified status}
\usage{
list_jobs_by_status(Status, Ascending = NULL, PageToken = NULL)
}
\arguments{
\item{Status}{[required] To get information about all of the jobs associated with the current AWS account that have a given status, specify the following status: \code{Submitted}, \code{Progressing}, \code{Complete}, \code{Canceled}, or \code{Error}.}

\item{Ascending}{To list jobs in chronological order by the date and time that they were submitted, enter \code{true}. To list jobs in reverse chronological order, enter \code{false}.}

\item{PageToken}{When Elastic Transcoder returns more than one page of results, use \code{pageToken} in subsequent \code{GET} requests to get each successive page of results.}
}
\description{
The ListJobsByStatus operation gets a list of jobs that have a specified status. The response body contains one element for each job that satisfies the search criteria.
}
\section{Accepted Parameters}{
\preformatted{list_jobs_by_status(
  Status = "string",
  Ascending = "string",
  PageToken = "string"
)
}
}

