% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.elastictranscoder_operations.R
\name{list_pipelines}
\alias{list_pipelines}
\title{The ListPipelines operation gets a list of the pipelines associated with the current AWS account}
\usage{
list_pipelines(Ascending = NULL, PageToken = NULL)
}
\arguments{
\item{Ascending}{To list pipelines in chronological order by the date and time that they were created, enter \code{true}. To list pipelines in reverse chronological order, enter \code{false}.}

\item{PageToken}{When Elastic Transcoder returns more than one page of results, use \code{pageToken} in subsequent \code{GET} requests to get each successive page of results.}
}
\description{
The ListPipelines operation gets a list of the pipelines associated with the current AWS account.
}
\section{Accepted Parameters}{
\preformatted{list_pipelines(
  Ascending = "string",
  PageToken = "string"
)
}
}

