% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.elastictranscoder_package.R
\docType{package}
\name{paws.elastictranscoder-package}
\alias{paws.elastictranscoder}
\alias{paws.elastictranscoder-package}
\title{paws.elastictranscoder: Amazon Elastic Transcoder}
\description{
The AWS Elastic Transcoder Service.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker
}

}
\keyword{internal}
