% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.elastictranscoder_operations.R
\name{read_job}
\alias{read_job}
\title{The ReadJob operation returns detailed information about a job}
\usage{
read_job(Id)
}
\arguments{
\item{Id}{The identifier of the job for which you want to get detailed information.}
}
\description{
The ReadJob operation returns detailed information about a job.
}
\examples{

}
