% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.elastictranscoder_operations.R
\name{read_preset}
\alias{read_preset}
\title{The ReadPreset operation gets detailed information about a preset}
\usage{
read_preset(Id)
}
\arguments{
\item{Id}{[required] The identifier of the preset for which you want to get detailed information.}
}
\description{
The ReadPreset operation gets detailed information about a preset.
}
\section{Accepted Parameters}{
\preformatted{read_preset(
  Id = "string"
)
}
}

