% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.elastictranscoder_operations.R
\name{update_pipeline_status}
\alias{update_pipeline_status}
\title{The UpdatePipelineStatus operation pauses or reactivates a pipeline, so that the pipeline stops or restarts the processing of jobs}
\usage{
update_pipeline_status(
  Id = "string",
  Status = "string"
)
}
\arguments{
\item{Id}{[required] The identifier of the pipeline to update.}

\item{Status}{[required] The desired status of the pipeline:
\itemize{
\item \code{Active}: The pipeline is processing jobs.
\item \code{Paused}: The pipeline is not currently processing jobs.
}}
}
\description{
The UpdatePipelineStatus operation pauses or reactivates a pipeline, so that the pipeline stops or restarts the processing of jobs.
}
\details{
Changing the pipeline status is useful if you want to cancel one or more jobs. You can't cancel jobs after Elastic Transcoder has started processing them; if you pause the pipeline to which you submitted the jobs, you have more time to get the job IDs for the jobs that you want to cancel, and to send a CancelJob request.
}
\examples{

}
