% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.elastictranscoder_operations.R
\name{read_pipeline}
\alias{read_pipeline}
\title{The ReadPipeline operation gets detailed information about a pipeline}
\usage{
read_pipeline(Id)
}
\arguments{
\item{Id}{[required] The identifier of the pipeline to read.}
}
\description{
The ReadPipeline operation gets detailed information about a pipeline.
}
\section{Accepted Parameters}{
\preformatted{read_pipeline(
  Id = "string"
)
}
}

